fix(shared): move test using files to e2e test
