+++ shared/src/commands/validate/validate.e2e.spec.ts
@@ -1,20 +1,20 @@
 import { describe, it, expect } from 'vitest';
-import { validate } from './validate.js';
+import { validate, applyArchitectureOptionsToPattern } from './validate.js';
 import { readFileSync } from 'fs';
-import { join } from 'path';
+import path from 'path';
 import { FileSystemDocumentLoader } from '../../document-loader/file-system-document-loader.js';
 import { SchemaDirectory } from '../../schema-directory.js';
 
-const inputArchPath = join(
+const inputArchPath = path.join(
     __dirname,
     '../../../test_fixtures/command/validate/options/arch.json'
 );
-const inputPatternPath = join(
+const inputPatternPath = path.join(
     __dirname,
     '../../../test_fixtures/command/validate/options/pattern.json'
 );
 
-const schemaDir = join(__dirname, '../../../../calm/release/1.0/meta/');
+const schemaDir = path.join(__dirname, '../../../../calm/release/1.0/meta/');
 
 describe('validate E2E', () => {
     let schemaDirectory;
@@ -36,5 +36,22 @@ describe('validate E2E', () => {
         expect(response.spectralSchemaValidationOutputs).toHaveLength(2);
         expect(response.spectralSchemaValidationOutputs[0].path).toBe('/nodes/0/description');
         expect(response.spectralSchemaValidationOutputs[1].path).toBe('/nodes/1/description');
+    });
+
+    describe('applyArchitectureOptionsToPattern', () => {
+        it('works with one options relationship', async () => {
+            const architecture = JSON.parse(
+                readFileSync(inputArchPath, 'utf8')
+            );
+            const pattern = JSON.parse(
+                readFileSync(inputPatternPath, 'utf8')
+            );
+            const expectedResult = JSON.parse(
+                readFileSync(path.join(__dirname, '../../../test_fixtures/command/validate/options/pattern-resolved.json'), 'utf8')
+            );
+
+            const newPattern = applyArchitectureOptionsToPattern(architecture, pattern);
+            expect(newPattern).toStrictEqual(expectedResult);
+        });
     });
 });
